+++ packages/frontend/src/scripts/get-bg-color.ts
